+++ cableprobe/models.py
@@ -45,6 +45,7 @@ KIND_KERNEL_MODULE = "kernel_module"
 KIND_MOUNT = "mount"
 KIND_WIFI_AP = "wifi_ap"
 KIND_KEYSTROKE_TIMING = "keystroke_timing"
+KIND_POWER_READING = "power_reading"
 KIND_PROCESS = "process"
 KIND_KERNEL_MESSAGE = "kernel_message"